refactor(agent_meta): extract transcript/config_files/resources helpers
Three submodules extracted in one pass; each kept well under the 512-line
hook ceiling and re-exported by agent_meta to preserve am._foo access:

  _meta/transcript.py    - _encode_claude_project, _latest_jsonls,
                           _read_sdk_session_state
  _meta/config_files.py  - _config_candidates, _read_claude_md,
                           _redact_mcp_tree, _read_mcp_json,
                           _parse_mcp_servers
  _meta/resources.py     - _pids_from_session, plus the inline psutil
                           block lifted to _collect_host_metrics()

Body of collect_rich now calls _collect_host_metrics() in place of the
40-line inline block. Output dict identical: psutil branch still returns
the same 14 keys with the same rounded values; empty {} on failure.

Integer literals in moved code use PEP-515 separators (20_000, 10_000,
and a _MB = 1024*1024 helper for the byte->MB math) — formatting only,
arithmetic identical.

412 _state/ tests pass.
